Develop iter_section to parse Sphinx sections
The idea behind iter_sections is to break a Sphinx-generated HTML pages
into hierarchically-sectioned chunks of content. Each Section object
keeps track of the section hierarchy, while also retaining the content
that is found within the section (exclusive of subsections).

The Section data structure will be the basis of records that are sent to
Algolia.