Add the TutorialSectionRecord
"Records" are types that are made to be turned directly into Algolia
index records. In this first case, the TutorialSectionRecord wraps a
Section instance of a tutorial page.

The object_id parameter is designed to create a unique ID given the URL
and section of the tutorial.

The output data structure is based on
https://blog.algolia.com/how-to-build-a-helpful-search-for-technical-documentation-the-laravel-example/#pitfall-nº1%C2%A0the-web-page-as-the-default%C2%A0entry