Implement index_tutorial workflow
The idea of workflows is that they're stateless, asynchronous functions
that process a unit of data to support astropy-librarian's indexing
operations.

In particular, index_tutorial takes a tutorial's URL and outputs records
to Algolia.

Supporting this is also the download_url workflow function, which might
be reused by other indexing operations.

This is hooked up into the index-tutorial demo script.