GitHub Actions CI now uses tox
This splits the CI up so that linting and type checking are separate
from running unit tests; unit tests now run the matrix of versions.

For linting, we use the pre-commit action because it's already set up to
cache dependencies. For the other tox environments we cache using
actions/cache. The hashing on setup.cfg, pyproject.toml, and tox.ini
are all necessary because they affect the tox environment.