Add test files from ccd-guide
This test dataset is from [A guide to CCD data reduction and stellar
photometry using astropy and affiliated
packages](https://github.com/astropy/ccd-reduction-and-photometry-guide)
and is used to test ingesting a Jupyter Book-rendered guide.

The index.html is used to identify the redirect page that Jupyter Book
creates.

The 00-00-Preface.html file is used to test parsing the homepage and the
<nav> for a list of all content in the guide.

The 01-05-Calibration-overview.html file is an example of a content page
that has a mix of text, math, code, and images.