Refactor download_html to return HtmlPage
The new HtmlPage class provides context and utilities around a
downloaded HTML page, such as the URL that the client was redirected to
in additional to the original URL, and any headers that went along with
the response. Additionally, HtmlPage also hosts the parser
functionality.

Finally, this merges with the TestData class in conftest.py, providing a
more unified way of passing around info about an HTML page between both
the application code and mocked up tests.