Create an Algolia index context manager
This context manager does a few things for us:

1. Rather than needing to pass around both the SearchClient and the
index name and run init_index on the spot, this context manager allows
us to centrally set up the index.

2. It also provides a layer of indirection over the use of Algolia,
enabling us to create a mock index to use for dry-run operations (this
is still experimental; index methods need to be added to the mock as
they come up).