Omit cell_output from ingest
This is a decision we might want to re-address, but for this this commit
effectively omits Jupyter output cells from the Algolia index. The
technical reason driving this is that output cells can contain *a lot*
of content, particularly if its a large table. Algolia needs to record
size to be <= 10KB, so omitting outputs helps us stay within that budget
without doing extra work.

Another reason is that arguably the output is less relevant for
searches. People are more likely to be searching for terms found in
either prose or the code input cells.