Implement object expiration with index_epoch key
At the end of an indexing workflow, now run the expire_old_records
workflow that finds records for the same root_url, but a different
index_epoch, and deletes them.

In most cases, objects are replaced because object IDs are predictable
based on the url and heading. But if a URL or heading is
changed/deleted, these object IDs won't be replaced by new content, and
instead need to be explicitly deleted by this workflow.