Add a priority field to records
Before a user starts to enter a search term on the Learn Astropy
homepage, we want the search results to show records, with curated
resources appearing first. How we can achieve that is by adding a
"priority" field to the Algolia records. By default, records have a
priority of 0. But by setting a --priority option to the index commands,
this priority can be elevated. Higher priorities appear first.