Split records for Algolia size limit
The maximum record size in the free plan is 10KB, so we need to split
records that are larger than that (i.e., by splitting their content
field). The new export_capped_records_to_algolia method for
AlgoliaRecord accomplishes that this calculates the
size of the JSON playload, and then uses the chunked algorithm from
more-itertools to create additional records with smaller chunks of the
original record's content.